yatube_api/requirements.txt
# Django & utils
Django==2.2.16
Pillow==8.3.1
requests==2.26.0
# pytest
pytest==6.2.4
pytest-pythonpath==0.7.3
pytest-django==4.4.0
# api
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
PyJWT==2.1.0
# lintering
flake8==5.0.4
flake8-broken-line==0.6.0
flake8-isort==5.0.0
flake8-plugin-utils==1.3.2
flake8-return==1.2.0
# third-party modules
ipython==7.34.0
python-dotenv==0.21.1

# Django productions
gunicorn~=20.1.0
whitenoise~=6.4.0
psycopg2-binary~=2.9.5
